+++ src/visits/charts/LineChartCard.tsx
@@ -1,5 +1,17 @@
+import {
+  faDownLeftAndUpRightToCenter as collapseIcon,
+  faUpRightAndDownLeftFromCenter as expandIcon,
+} from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { countBy } from '@shlinkio/data-manipulation';
-import { HIGHLIGHTED_COLOR, isDarkThemeEnabled, MAIN_COLOR } from '@shlinkio/shlink-frontend-kit';
+import {
+  HIGHLIGHTED_COLOR,
+  isDarkThemeEnabled,
+  MAIN_COLOR,
+  useElementRef,
+  useToggle,
+} from '@shlinkio/shlink-frontend-kit';
+import { clsx } from 'clsx';
 import type { Duration } from 'date-fns';
 import {
   add,
@@ -15,8 +27,11 @@ import {
   startOfISOWeek,
 } from 'date-fns';
 import type { FC } from 'react';
+import { useId } from 'react';
+import { useLayoutEffect } from 'react';
 import { Fragment, useCallback, useMemo, useState } from 'react';
 import {
+  Button,
   Card,
   CardBody,
   CardHeader,
@@ -30,7 +45,7 @@ import type { CategoricalChartState } from 'recharts/types/chart/types';
 import { formatInternational } from '../../utils/dates/helpers/date';
 import type { StrictDateRange } from '../../utils/dates/helpers/dateIntervals';
 import { rangeOf } from '../../utils/helpers';
-import { useMaxResolution } from '../../utils/helpers/hooks';
+import { useKeyDown, useMaxResolution } from '../../utils/helpers/hooks';
 import { prettify } from '../../utils/helpers/numbers';
 import type { MediaMatcher } from '../../utils/types';
 import type { NormalizedVisit, Stats } from '../types';
@@ -199,7 +214,7 @@ const useActiveDot = (
 };
 
 const payloadFromChartEvent = (e: CategoricalChartState) =>
-  e.activePayload?.[0].payload as ChartPayloadEntry | undefined;
+  e.activePayload?.[0]?.payload as ChartPayloadEntry | undefined;
 
 export type LineChartCardProps = {
   visitsGroups: Record<string, VisitsList>;
@@ -234,11 +249,39 @@ export const LineChartCard: FC<LineChartCardProps> = (
   }, [step, visitsGroups]);
   const activeDot = useActiveDot(visitsGroups, step, setSelectedVisits);
 
+  const [isExpanded, toggleExpanded,, setNotExpanded] = useToggle();
+  const bodyRef = useElementRef<HTMLElement>();
+  const bodyId = useId();
+  const legendRef = useElementRef<HTMLUListElement>();
+  const [wrapperHeight, setWrapperHeight] = useState(isMobile ? 300 : 400);
+
+  useLayoutEffect(() => {
+    if (!isExpanded) {
+      setWrapperHeight(isMobile ? 300 : 400);
+      return () => {};
+    }
+
+    const observer = new ResizeObserver(() => {
+      const { height: bodyHeight } = bodyRef.current!.getBoundingClientRect();
+      const { height: legendHeight } = legendRef.current!.getBoundingClientRect();
+
+      // 32 is the body's padding, 16 is the legend's top margin
+      const offset = 32 + 16;
+
+      setWrapperHeight(bodyHeight - legendHeight - offset);
+    });
+    observer.observe(bodyRef.current!);
+    observer.observe(legendRef.current!);
+
+    return () => observer.disconnect();
+  }, [bodyRef, isExpanded, isMobile, legendRef]);
+
+  useKeyDown('Escape', setNotExpanded, isExpanded);
+
   const ChartWrapper = dimensions ? Fragment : ResponsiveContainer;
-  const wrapperDimensions = useMemo(
-    // If dimensions were explicitly provided for the chart, we don't need to set dimensions in the wrapper as well
-    () => (dimensions ? {} : { width: '100%', height: isMobile ? 300 : 400 }),
-    [dimensions, isMobile],
+  const wrapperProps = useMemo(
+    () => (dimensions ? {} : { width: '100%', height: wrapperHeight }),
+    [dimensions, wrapperHeight],
   );
 
   // References the items being selected via drag'n'drop
@@ -275,24 +318,36 @@ export const LineChartCard: FC<LineChartCardProps> = (
   }, [onDateRangeChange, resetSelection, selectionEnd, selectionStart]);
 
   return (
-    <Card>
+    <Card className={clsx({ 'fixed-top fixed-bottom': isExpanded })} data-testid="line-chart-card">
       <CardHeader role="heading" aria-level={4} className="d-flex justify-content-between align-items-center">
         Visits over time
-        <UncontrolledDropdown>
-          <DropdownToggle caret color="link" className="btn-sm p-0">
-            Group by
-          </DropdownToggle>
-          <DropdownMenu end>
-            {Object.entries(STEPS_MAP).map(([value, menuText]) => (
-              <DropdownItem key={value} active={step === value} onClick={() => setStep(value as Step)}>
-                {menuText}
-              </DropdownItem>
-            ))}
-          </DropdownMenu>
-        </UncontrolledDropdown>
+        <div className="d-flex align-content-center gap-1">
+          <Button
+            aria-label={isExpanded ? 'Collapse' : 'Expand'}
+            aria-expanded={isExpanded}
+            aria-controls={bodyId}
+            size="sm"
+            color="link"
+            onClick={toggleExpanded}
+          >
+            <FontAwesomeIcon icon={isExpanded ? collapseIcon : expandIcon} />
+          </Button>
+          <UncontrolledDropdown className="d-flex align-items-center">
+            <DropdownToggle caret color="link" className="btn-sm p-0">
+              Group by
+            </DropdownToggle>
+            <DropdownMenu end>
+              {Object.entries(STEPS_MAP).map(([value, menuText]) => (
+                <DropdownItem key={value} active={step === value} onClick={() => setStep(value as Step)}>
+                  {menuText}
+                </DropdownItem>
+              ))}
+            </DropdownMenu>
+          </UncontrolledDropdown>
+        </div>
       </CardHeader>
-      <CardBody>
-        <ChartWrapper {...wrapperDimensions}>
+      <CardBody innerRef={bodyRef} id={bodyId}>
+        <ChartWrapper {...wrapperProps}>
           <LineChart
             className="user-select-none"
             data={chartData}
@@ -323,7 +378,7 @@ export const LineChartCard: FC<LineChartCardProps> = (
             )}
           </LineChart>
         </ChartWrapper>
-        <LineChartLegend visitsGroups={visitsGroups} />
+        <LineChartLegend visitsGroups={visitsGroups} ref={legendRef} />
       </CardBody>
     </Card>
   );